fix to albumentations

+++ src/data/project2/dataloader.py
@@ -138,8 +138,8 @@ class DRIVE(torch.utils.data.Dataset):
         image = np.array(Image.open(image_path))
         label =  np.array(Image.open(label_path)) * 1
         transformed = self.transform(image=image, mask=label)
-        X = transformed['image']
-        Y = transformed['mask']
+        X = transformed['image'].float()
+        Y = transformed['mask'].float()
         return X, Y
     
 ## Dataset classes - PH2
@@ -185,6 +185,6 @@ class PH2_dataset(torch.utils.data.Dataset):
         image = np.array(Image.open(image_path))
         label =  np.array(Image.open(label_path)) * 1
         transformed = self.transform(image=image, mask=label)
-        X = transformed['image']
-        Y = transformed['mask']
+        X = transformed['image'].float()
+        Y = transformed['mask'].float()
         return X, Y
